+++ headers/layout.h
@@ -4,9 +4,9 @@
 #include "graph.h"
 
 const int ITERATION = 300;
-const float K = 0.1;
-const float COOLING_FACTOR = 0.8;
-const float E = 2 * NODE_RADIUS; // minimum distance
+const double K = 0.1;
+const double COOLING_FACTOR = 0.8;
+const double E = 2 * NODE_RADIUS; // minimum distance
 
 class Layout {
 public:
@@ -19,11 +19,11 @@ public:
 
 private:
   void fruchtermanReingold(const std::unique_ptr<Graph> &graph, int iterations,
-                           float k, double width, double height, int x, int y);
+                        double k, int width, int height, int x, int y);
   void layoutBruteForce(const std::unique_ptr<Graph> &graph, int width,
                         int height, int iterations, int x, int y);
 
-  static float distance(float dx, float dy);
+  static double distance(double dx, double dy);
 };
 
 #endif
